refactor(agent): 抽取共享数据类型到 workflow/types 基础包
